requirements.txt
numpy==2.3.5
pandas==2.3.3
requests==2.32.5
matplotlib==3.10.7
pillow==12.0.0
faker==38.2.0
python-dateutil==2.9.0
rich==13.9+*
pyfiglet==1.0.4
emoji==2.14+
